update group list
