+++ lib/data/group_data.dart
@@ -1,0 +1,23 @@
+
+final List<Map<String, dynamic>> groups = [
+  {
+    'name': 'Petani Lama',
+    'memberCount': 189,
+    'imageUrl': 'assets/images/group1.png', // Replace with your image path
+  },
+  {
+    'name': 'Hobi Bercucuk Tanam',
+    'memberCount': 234,
+    'imageUrl': 'assets/images/group2.png', // Replace with your image path
+  },
+  {
+    'name': 'Benih Sihat Murah',
+    'memberCount': 240,
+    'imageUrl': 'assets/images/group3.png', // Replace with your image path
+  },
+  {
+    'name': 'Ladang di Melaka',
+    'memberCount': 145,
+    'imageUrl': 'assets/images/group4.png', // Replace with your image path
+  },
+];
